requirements.txt
matplotlib==3.5.1
numpy==1.21.5
opencv_contrib_python==4.6.0.66
opencv_python==4.5.5.64
Pillow==6.1.0
Pillow==8.2.0
Pillow==11.0.0
pytorch_wavelets==1.3.0
scikit_learn==1.0.2
timm==0.5.4
torch==1.12.1+cu116
torchvision==0.13.1+cu116
